SF Bugs#99: Increase limit for Map IDs
Change from a guint8 to a guint16 to enable larger map id range.
Thus prevents silently auto wrapping numbers around which may then overwrite a different map cache.

Conflicts:
	src/vikslippymapsource.h

+++ src/mapcache.h
@@ -28,9 +28,9 @@
 G_BEGIN_DECLS
 
 void a_mapcache_init ();
-void a_mapcache_add ( GdkPixbuf *pixbuf, gint x, gint y, gint z, guint8 type, guint zoom, guint8 alpha, gdouble xshrinkfactor, gdouble yshrinkfactor );
-GdkPixbuf *a_mapcache_get ( gint x, gint y, gint z, guint8 type, guint zoom, guint8 alpha, gdouble xshrinkfactor, gdouble yshrinkfactor );
-void a_mapcache_remove_all_shrinkfactors ( gint x, gint y, gint z, guint8 type, guint zoom );
+void a_mapcache_add ( GdkPixbuf *pixbuf, gint x, gint y, gint z, guint16 type, guint zoom, guint8 alpha, gdouble xshrinkfactor, gdouble yshrinkfactor );
+GdkPixbuf *a_mapcache_get ( gint x, gint y, gint z, guint16 type, guint zoom, guint8 alpha, gdouble xshrinkfactor, gdouble yshrinkfactor );
+void a_mapcache_remove_all_shrinkfactors ( gint x, gint y, gint z, guint16 type, guint zoom );
 void a_mapcache_flush ();
 void a_mapcache_uninit ();
