+++ src/vikwmscmapsource.h
@@ -48,7 +48,7 @@ struct _VikWmscMapSource
 
 GType vik_wmsc_map_source_get_type (void) G_GNUC_CONST;
 
-VikWmscMapSource * vik_wmsc_map_source_new_with_id (guint8 id, const gchar *label, const gchar *hostname, const gchar *url);
+VikWmscMapSource * vik_wmsc_map_source_new_with_id (guint16 id, const gchar *label, const gchar *hostname, const gchar *url);
 
 G_END_DECLS
